requirements.txt
argostranslate==1.9.2 
grpcio==1.62.1
grpcio-tools==1.62.1
python-dotenv==1.0.1
black==24.2.0
